project

+++ src/by/it/obmetko/project/java/commands/crud/CommandCrudCategories.java
@@ -3,10 +3,7 @@ package by.it.obmetko.project.java.commands.crud;
 import by.it.obmetko.project.java.DAO.DAO;
 import by.it.obmetko.project.java.DAO.beens.Catalog;
 import by.it.obmetko.project.java.DAO.beens.Category;
-import by.it.obmetko.project.java.controller.Action;
-import by.it.obmetko.project.java.controller.ActionCommand;
-import by.it.obmetko.project.java.controller.FormUtil;
-import by.it.obmetko.project.java.controller.Msg;
+import by.it.obmetko.project.java.controller.*;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -14,32 +11,30 @@ import java.sql.SQLException;
 import java.text.ParseException;
 import java.util.List;
 
-
-
-
-
-
-
-
 public class CommandCrudCategories extends ActionCommand {
 
     @Override
-    public ActionCommand execute(HttpServletRequest req, HttpServletResponse resp) throws ParseException, SQLException {        if (FormUtil.isPost(req)) {
+    public ActionCommand execute(HttpServletRequest req, HttpServletResponse resp) throws ParseException, SQLException {
+        if (FormUtil.isPost(req)) {
             Category category=new Category(FormUtil.getInt(req.getParameter("id")),
-                    FormUtil.getString(req.getParameter("email"), req.getParameter("categoryName")),
-                    FormUtil.getInt(req.getParameter("Catalog_id")));
-            if(req.getParameter("Update")!=null){
+                                       FormUtil.getString(req.getParameter("categoryName"),Pattern.CATEGORY),
+                                        FormUtil.getInt(req.getParameter("Catalog_id")));
+
+
+            if (req.getParameter("Update") != null) {
                 DAO.getDAO().categoryDAO.update(category);
-            }else if(req.getParameter("Delete")!=null){
+            } else if (req.getParameter("Delete") != null) {
                 DAO.getDAO().categoryDAO.delete(category);
-            }else if(req.getParameter("Create")!=null){
+            } else if (req.getParameter("Create") != null) {
                 DAO.getDAO().categoryDAO.create(category);
             }
         }
+
+
         List<Category> categories = DAO.getDAO().categoryDAO.getAll("");
         List<Catalog> catalogs = DAO.getDAO().catalogDAO.getAll("");
         req.setAttribute(Msg.CATEGORIES, categories);
         req.setAttribute(Msg.CATALOGS, catalogs);
-        return Action.CRUDCATEGORIES.command;
+        return Actions.CRUDCATEGORIES.command;
     }
 }
